Combine helpers-nodes and helpers-results.
